feat: moved minimap area canvas to pinia store
