+++ src/store/MinimapareaCanvas/getters.ts
@@ -1,0 +1,1 @@
+export const getters = {}
