+++ src/store/MinimapareaCanvas/states.ts
@@ -1,0 +1,12 @@
+export const states = {
+    canvas: document.getElementById('miniMapArea'),
+    ctx: {},
+    pageHeight: 0,
+    pageWidth: 0,
+    pageY: 0,
+    pageX: 0,
+    minY: 0,
+    maxY: 0,
+    minX: 0,
+    maxX: 0
+}
